backend/composite-services/team-swap/requirements.txt
flask
flask_sqlalchemy
marshmallow
python-dotenv
sqlalchemy
psycopg2-binary
ortools
flask-openapi3
pydantic
requests

